Fix environmental problems.
Environment is now a list of lists, not just a list. It has multiple
frames as alists. So the current environments are:

env -+- frame 1 (the current)
     |
     +- frame 2 (1-level outer)
     :
     :
     +- frame top (the only frame for the top-level environment)

And the frames (alists) are treated as destructive containers so that
the same Values mean the same frames. Environments (which made of
frames) need to be destructive because a captured environent will be
used after some changes (e.g. variable definitions) and that changes
need to be effective when the captured environment used.

For example,

    (define f (lambda () a))
    (define a 42)
    (f)

cannot raise error but evaluated as 42 even if the first (lambda)
captures a environment without a.